Implemented basics of scalars for getting data rate on digitizers. Still can't figure out how CoMPASS gets the board input rate, but at least have throughput

+++ src/Platform/OpenGL/OpenGLRendererAPI.cpp
@@ -14,8 +14,8 @@ namespace BoxScore {
 		glClearColor(color[0], color[1], color[2], color[3]);
 	}
 
-	float OpenGLRendererAPI::GetFrameTime()
+	double OpenGLRendererAPI::GetFrameTime()
 	{
-		return (float)glfwGetTime();
+		return glfwGetTime();
 	}
 }
